services/ride-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
psycopg2-binary
sqlalchemy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp